requirements.txt
fastapi
uvicorn
pytube
openai-whisper
imageio-ffmpeg